Added another version which scrapes via Ajax
The existing code worked for me, but I was unable to arm using force bypass. Rewrote the code to scrape and arm via the regular site using Ajax/JSON. The general idea is the same as the old code, but there are a few more steps because the relevant information is scattered across several URLs. This version can take in optional forcebypass and noentrydelay fields to pass along to the arming events. The sensor_status is not exactly the same as the pda version, but it should reflect similar information.
I just dropped this new version in a different file because I'm not sure how Daren wants to integrate it. We should probably keep both this and the pda version as the scraping of either version may break if ADC makes changes. We can let the user pick the scraping version to use via the HA config.